[Feature][Server] Add data quality report feature (#404)

+++ datavines-server/src/main/java/io/datavines/server/repository/service/impl/JobExecutionServiceImpl.java
@@ -16,6 +16,7 @@
  */
 package io.datavines.server.repository.service.impl;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
@@ -317,7 +318,7 @@ public class JobExecutionServiceImpl extends ServiceImpl<JobExecutionMapper, Job
             ResultFormula resultFormula =
                     PluginLoader.getPluginLoader(ResultFormula.class).getOrCreatePlugin(result.getResultFormula());
             MetricExecutionDashBoard executionDashBoard = new MetricExecutionDashBoard();
-            executionDashBoard.setValue(resultFormula.getResult(result.getActualValue(), Objects.isNull(result.getExpectedValue()) ? 0 : result.getExpectedValue()));
+            executionDashBoard.setValue(resultFormula.getResult(result.getActualValue(), Objects.isNull(result.getExpectedValue()) ? BigDecimal.valueOf(0) : result.getExpectedValue()));
             executionDashBoard.setType(resultFormula.getType().getDescription());
             executionDashBoard.setDatetime(result.getCreateTime().toString());
